Update to bsnes v048 release.
The biggest feature of this new release is the addition of save state support. Note that this is only currently supported for normal games, and the SPC7110 and OBC-1 co-processors. Other special chips, such as the SuperFX and SA-1, cannot currently save and load state files. I will be adding support for other co-processors little by little in future releases.
Changelog:
    - Added save state support
    - Added SPC7110 and OBC1 save state support
    - Added new tools group, with new cheat code and save state managers
    - Lots of new UI shortcuts: quick save state, quick load state, show state manager, etc
    - Escape key will now close both the settings and tools group windows
    - Added major speed-ups to both SuperFX and SA-1 emulation; both now run ~15-25% faster than v047
    - Added new video filter, LQ2x; it's as fast as Scale2x while being almost as smooth as HQ2x
    - Re-wrote HQ2x algorithm; code size was reduced to less than 10% of its original size with virtually no speed loss
    - Corrected SuperFX2 cache access timing; fixes Stunt Race FX menus and slowdown in other titles
    - Relaxed palette write limitations for PGA Tour Golf [Jonas Quinn]
    - Fixed a slight timing issue that was breaking 'An Americal Tail - Feivel Goes West'
    - Turned off auto-save of SRAM as it was causing slowdowns when writing to flash memory; can be re-enabled via bsnes.cfg -> system.autoSaveMemory = true
    - Added bsnes.cfg -> system.autoHideMenus, defaults to false; when true, menu and status bars will be hidden upon entering fullscreen mode
    - Added skeletons for ST011 and ST018 support. Both Quick-move titles get in-game now
    - Re-wrote S-CPU and S-SMP processor cores to use templates, removed custom pre-processor
    - Split PPUcounter into a base class inherited by both PPU and CPU; allows both cores to run out-of-order
    - Split inline header functions to separate files, allows headers to be included in any order now

+++ src/chip/superfx/core/registers.hpp
@@ -1,12 +1,10 @@
-
-
 //accepts a callback binding so r14 writes can trigger ROM buffering transparently
 struct reg16_t : noncopyable {
-  uint16_t data;
-  function<void (uint16_t)> on_modify;
+  uint16 data;
+  function<void (uint16)> on_modify;
 
   inline operator unsigned() const { return data; }
-  inline uint16_t assign(uint16_t i) {
+  inline uint16 assign(uint16 i) {
     if(on_modify) on_modify(i);
     else data = i;
     return data;
@@ -47,7 +45,7 @@ struct sfr_t {
   bool cy;    //carry flag
   bool z;     //zero flag
 
-  operator uint16_t() const {
+  operator unsigned() const {
     return (irq << 15) | (b << 12) | (ih << 11) | (il << 10) | (alt2 << 9) | (alt1 << 8)
          | (r << 6) | (g << 5) | (ov << 4) | (s << 3) | (cy << 2) | (z << 1);
   }
@@ -75,11 +73,11 @@ struct scmr_t {
   bool ran;
   unsigned md;
 
-  operator uint8_t() const {
+  operator unsigned() const {
     return ((ht >> 1) << 5) | (ron << 4) | (ran << 3) | ((ht & 1) << 2) | (md);
   }
 
-  scmr_t& operator=(uint8_t data) {
+  scmr_t& operator=(uint8 data) {
     ht  = (bool)(data & 0x20) << 1;
     ht |= (bool)(data & 0x04) << 0;
     ron = data & 0x10;
@@ -96,11 +94,11 @@ struct por_t {
   bool dither;
   bool transparent;
 
-  operator uint8_t() const {
+  operator unsigned() const {
     return (obj << 4) | (freezehigh << 3) | (highnibble << 2) | (dither << 1) | (transparent);
   }
 
-  por_t& operator=(uint8_t data) {
+  por_t& operator=(uint8 data) {
     obj         = data & 0x10;
     freezehigh  = data & 0x08;
     highnibble  = data & 0x04;
@@ -114,11 +112,11 @@ struct cfgr_t {
   bool irq;
   bool ms0;
 
-  operator uint8_t() const {
+  operator unsigned() const {
     return (irq << 7) | (ms0 << 5);
   }
 
-  cfgr_t& operator=(uint8_t data) {
+  cfgr_t& operator=(uint8 data) {
     irq = data & 0x80;
     ms0 = data & 0x20;
     return *this;
@@ -126,30 +124,30 @@ struct cfgr_t {
 };
 
 struct regs_t {
-  uint8_t pipeline;
-  uint16_t ramaddr;
+  uint8 pipeline;
+  uint16 ramaddr;
 
   reg16_t r[16];    //general purpose registers
   sfr_t sfr;        //status flag register
-  uint8_t pbr;      //program bank register
-  uint8_t rombr;    //game pack ROM bank register
+  uint8 pbr;        //program bank register
+  uint8 rombr;      //game pack ROM bank register
   bool rambr;       //game pack RAM bank register
-  uint16_t cbr;     //cache base register
-  uint8_t scbr;     //screen base register
+  uint16 cbr;       //cache base register
+  uint8 scbr;       //screen base register
   scmr_t scmr;      //screen mode register
-  uint8_t colr;     //color register
+  uint8 colr;       //color register
   por_t por;        //plot option register
   bool bramr;       //back-up RAM register
-  uint8_t vcr;      //version code register
+  uint8 vcr;        //version code register
   cfgr_t cfgr;      //config register
   bool clsr;        //clock select register
 
   unsigned romcl;   //clock ticks until romdr is valid
-  uint8_t romdr;    //ROM buffer data register
+  uint8 romdr;      //ROM buffer data register
 
   unsigned ramcl;   //clock ticks until ramdr is valid
-  uint16_t ramar;   //RAM buffer address register
-  uint8_t ramdr;    //RAM buffer data register
+  uint16 ramar;     //RAM buffer address register
+  uint8 ramdr;      //RAM buffer data register
 
   reg16_t *sreg, *dreg;
   reg16_t& sr() { return *sreg; }  //source register (from)
@@ -166,12 +164,12 @@ struct regs_t {
 } regs;
 
 struct cache_t {
-  uint8_t buffer[512];
+  uint8 buffer[512];
   bool valid[32];
 } cache;
 
 struct pixelcache_t {
-  uint16_t offset;
-  uint8_t bitpend;
-  uint8_t data[8];
+  uint16 offset;
+  uint8 bitpend;
+  uint8 data[8];
 } pixelcache[2];
